Fix the use of Vim built-in make command.
This is basically a hack. Since there is no other way to escape special
characters in filenames except using `fnameescape`, we have to
specifically set `makeprg` according to the filename. And in order to
catch all the rename, new file, open file cases, an autocmd is
introduced. Note that backward compatibility is preserved: command
`CoffeeMake` still exists and should do whatever what it was doing before.
But function `CoffeeMake` has been removed. Also the built-in `make` does
not work the same way as `CoffeeMake`, the difference is that `CoffeeMake`
is chainable while the built-in make is not.